PDJB-1039: move files to shared, and fix uneccesary code
