+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/journeys/propertyRegistration/states/InviteJointLandlordPropertyRegistrationState.kt
@@ -2,7 +2,8 @@ package uk.gov.communities.prsdb.webapp.journeys.propertyRegistration.states
 
 import uk.gov.communities.prsdb.webapp.journeys.propertyRegistration.steps.HasAnyJointLandlordsInvitedStep
 import uk.gov.communities.prsdb.webapp.journeys.propertyRegistration.steps.HasJointLandlordsStep
-import uk.gov.communities.prsdb.webapp.journeys.propertyRegistration.tasks.SharedInviteJointLandlordsTask
+import uk.gov.communities.prsdb.webapp.journeys.shared.inviteJointLandlord.SharedInviteJointLandlordsTask
+import uk.gov.communities.prsdb.webapp.journeys.shared.states.SharedInviteJointLandlordState
 
 interface InviteJointLandlordPropertyRegistrationState : SharedInviteJointLandlordState {
     val hasAnyJointLandlordsInvitedStep: HasAnyJointLandlordsInvitedStep
