+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/journeys/shared/states/SharedInviteJointLandlordState.kt
@@ -1,9 +1,9 @@
-package uk.gov.communities.prsdb.webapp.journeys.propertyRegistration.states
+package uk.gov.communities.prsdb.webapp.journeys.shared.states
 
 import uk.gov.communities.prsdb.webapp.journeys.JourneyState
-import uk.gov.communities.prsdb.webapp.journeys.propertyRegistration.steps.CheckJointLandlordsStep
-import uk.gov.communities.prsdb.webapp.journeys.propertyRegistration.steps.InviteJointLandlordStep
-import uk.gov.communities.prsdb.webapp.journeys.propertyRegistration.steps.RemoveJointLandlordAreYouSureStep
+import uk.gov.communities.prsdb.webapp.journeys.shared.inviteJointLandlord.CheckJointLandlordsStep
+import uk.gov.communities.prsdb.webapp.journeys.shared.inviteJointLandlord.InviteJointLandlordStep
+import uk.gov.communities.prsdb.webapp.journeys.shared.inviteJointLandlord.RemoveJointLandlordAreYouSureStep
 
 interface SharedInviteJointLandlordState : JourneyState {
     val inviteJointLandlordStep: InviteJointLandlordStep
