+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/pageObjects/pages/inviteJointLandlordJourneyPages/CheckJointLandlordsFormPageInviteJointLandlord.kt
@@ -8,7 +8,7 @@ import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.PostFo
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.SecondaryButton
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.SummaryList
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.basePages.BasePage
-import uk.gov.communities.prsdb.webapp.journeys.propertyRegistration.steps.CheckJointLandlordsStep
+import uk.gov.communities.prsdb.webapp.journeys.shared.inviteJointLandlord.CheckJointLandlordsStep
 
 class CheckJointLandlordsFormPageInviteJointLandlord(
     page: Page,
